requirements.txt
# requirements.txt
pynput
pyperclip
pillow